Refactor test participant communication mechanism:
Listen to receive events from the peer instead of having the router route test packets to the test participant.
Use the router to echo the packet and other info to the originating participant instead of sending to the test client directly via postMessage.
Simplify the test client implementation now that client callbacks are not used to forward test data.