feat: Integrate with latest Marketplace back-end
The API server root is now pointing at the server that provides marketplace data from the master branch. The Changes in
this commit address several changes in that data:

- Authentication is now required
- Link rel names are qualified with prefix ozp:
- Application listings are returned by specifying the appropriate HTTP Accept header

The resource keys for applications are now generated as /applicatin/{uuid}. Previously the key was
/application/{name}

The enpoints and Ajax classes were modified to enable setting of request headers when loading data from the
back-end server. The SystemApi data loading code now uses this to set the appropriate headers for retrieving
the desired representation of application listing. Loading of the APIs has not yet been set to configure the
appropriate Accept headers.
;